# Convert more cross joins to inner joins (Address performance/execution plan of TPCH query 19)