@avantgardnerio 2022-09-14
It might be important to note that master regressed in functionality: all the queries used to pass. Some were slow and only accurate to 4 decimals, but now several actually fail. Hopefully q19 isn't one of them.