@alamb 2022-09-14
>  It might be important to note that master regressed in functionality: all the queries used to pass. Some were slow and only accurate to 4 decimals, but now several actually fail. Hopefully q19 isn't one of them.

🤔  this sounds bad -- I thought we had tests in place to catch this. Clearly we didn't have enough.

I also plan to review this PR more carefully tomorrow