@avantgardnerio 2022-09-14
> this sounds bad

Ah, my bad - this may be behind a flag: https://github.com/apache/arrow-datafusion/pull/3393#issue-1364925083

One man's regression is another's more strict acceptance criteria.