@avantgardnerio 2022-09-14
> I think we should think of ways of adding the tpch queries to the CI?

There are tests for several of them here: https://github.com/apache/arrow-datafusion/blob/5f029cc73755b6800217e370ebe0a7b5e8a6a224/datafusion/core/tests/sql/subqueries.rs#L119