@kmitchener 2022-09-14
@Dandandan adding verification of TPCH results to the CI would be nice. I was thinking of suggesting that once all the tests pass and we can verify results against the answers. Fixing q19 is one of the last major blockers, besides some decimal issues.