@DhamoPS 2022-09-16
In predicates.rs, tpch_q19_pull_predicates_to_innerjoin_simplified() is having 3 tables test. I add one more test here as well.