@alamb 2022-09-17
To be clear, I think this PR just needs a few more tests, as discussed (and the clippy / fmt tests to be cleaned up) and it will be ready to merge

Thank you so much for your help @DhamoPS