@DhamoPS 2022-09-20
@alamb I have resolved the last clippy failure. I didn't know that I could run the clippy and lint tests locally. 
I am sorry for multiple commits. I will ensure that these tests are successful on my next PR.
Thanks a lot for your help!