@kmitchener 2022-09-20
This is great @DhamoPS! Congrats on your first merged PR :) Will you be following up with a separate PR to remove the similar code from sql/planner.rs?