@DhamoPS 2022-09-20
@kmitchener Thanks. Yes, I will be following up with separate PR for cleanup in sql/planner.rs.
Could you create the git issue to track it?