@kmitchener 2022-09-20
> @kmitchener Thanks. Yes, I will be following up with separate PR for cleanup in sql/planner.rs. Could you create the git issue to track it?

You should feel free to create an issue yourself when you're ready. I'm interested in the cleanup but I don't have any special ownership here.